Moved test from problem_helper.py to test_helpers.py
